refactor: refatorado model Players
